Use same optimize_theta as GP in SGP
